# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "SVT System", "description": "Secure Traceable Visual Token System \u2014 Production API", "version": "1.0.0"}, "paths": {"/api/v1/auth/register": {"post": {"tags": ["auth"], "summary": "Register", "description": "Register a new issuer account.", "operationId": "register_api_v1_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRegisterRequest"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRegisterResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "description": "Login and receive JWT tokens with device binding.", "operationId": "login_api_v1_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserLoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserLoginResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/refresh": {"post": {"tags": ["auth"], "summary": "Refresh", "description": "Refresh access token using refresh_token cookie.", "operationId": "refresh_api_v1_auth_refresh_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenRefreshResponse"}}}}}}}, "/api/v1/auth/logout": {"post": {"tags": ["auth"], "summary": "Logout", "description": "Logout: revoke refresh token by adding its jti to Redis.", "operationId": "logout_api_v1_auth_logout_post", "responses": {"204": {"description": "Successful Response"}}}}, "/api/v1/auth/scanner/create": {"post": {"tags": ["auth"], "summary": "Create Scanner Device", "description": "Create a scanner device (ISSUER or ADMIN only).\nReturns the API key once \u2014 it cannot be retrieved again.", "operationId": "create_scanner_device_api_v1_auth_scanner_create_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScannerDeviceCreateRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScannerDeviceCreateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/tokens": {"post": {"tags": ["tokens"], "summary": "Create Token", "description": "Generate an unbreakable cryptographically wrapped SVT representing a destination endpoint.\nEnforces strict idempotency per issuer.", "operationId": "create_token_api_v1_tokens_post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "title": "Idempotency-Key"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenCreateRequest"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["tokens"], "summary": "List Tokens", "description": "Fetch tokens owned strictly by the authenticated issuer context.\nProvides limit/offset pagination mechanism.", "operationId": "list_tokens_api_v1_tokens_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 20, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/TokenDetailsResponse"}, "title": "Response List Tokens Api V1 Tokens Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/tokens/{trace_id}/revoke": {"post": {"tags": ["tokens"], "summary": "Revoke Token", "description": "Revoke a specific SVT Token.", "operationId": "revoke_token_api_v1_tokens__trace_id__revoke_post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "trace_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Trace Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RevocationRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/tokens/admin/{trace_id}/unblock": {"post": {"tags": ["tokens"], "summary": "Unblock Token", "description": "Unblock a revoked SVT Token.", "operationId": "unblock_token_api_v1_tokens_admin__trace_id__unblock_post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "trace_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Trace Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/keys/rotate": {"post": {"tags": ["keys"], "summary": "Rotate Key", "description": "Rotate signing keys for the issuer.\nThis creates a new KeyVersion for the issuer without invalidating old ones.\nUsing FOR UPDATE lock to avoid race conditions.", "operationId": "rotate_key_api_v1_keys_rotate_post", "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Rotate Key Api V1 Keys Rotate Post"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/scan/verify": {"post": {"tags": ["scan"], "summary": "Verify Svt", "description": "High-velocity SVT verification endpoint.\nRate limited: 30 req/min per scanner_device_id.\nAuthenticated via HMAC-SHA256 scanner credentials.", "operationId": "verify_svt_api_v1_scan_verify_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScanVerifyRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScanVerifyResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/events/stream": {"get": {"tags": ["events"], "summary": "Sse Stream", "operationId": "sse_stream_api_v1_events_stream_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "Last-Event-Id", "in": "header", "required": false, "schema": {"type": "string", "default": "0-0", "title": "Last-Event-Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/health": {"get": {"tags": ["health"], "summary": "Liveness probe", "description": "Liveness probe \u2014 always returns OK if the process is running.\nUsed by Kubernetes liveness probe.", "operationId": "health_liveness_api_v1_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/health/ready": {"get": {"tags": ["health"], "summary": "Readiness probe", "description": "Readiness probe \u2014 validates DB and Redis connectivity.\nReturns 503 if either dependency is unreachable.\nUsed by Kubernetes readiness probe to gate traffic.", "operationId": "health_readiness_api_v1_health_ready_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "RevocationRequest": {"properties": {"reason": {"type": "string", "maxLength": 500, "minLength": 1, "title": "Reason"}}, "type": "object", "required": ["reason"], "title": "RevocationRequest"}, "ScanVerifyRequest": {"properties": {"svt_raw": {"type": "string", "title": "Svt Raw", "description": "The raw CBOR + Base64url envelope string parsed from the SVT."}, "device_fingerprint_override": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Device Fingerprint Override"}, "telemetry_lat": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Telemetry Lat"}, "telemetry_lng": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Telemetry Lng"}}, "type": "object", "required": ["svt_raw"], "title": "ScanVerifyRequest", "description": "Encapsulates the payload received from physically scanning the SVT."}, "ScanVerifyResponse": {"properties": {"result": {"type": "string", "enum": ["ALLOW", "DENY", "WARN", "CANNOT_VERIFY"], "title": "Result"}, "payload_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Payload Url", "description": "The authorized payload destination. Suppressed strictly on DENY."}, "reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reason", "description": "Contextual diagnostic string regarding the verdict."}}, "type": "object", "required": ["result"], "title": "ScanVerifyResponse", "description": "The normalized verdict routing directive transmitted back to the scanner."}, "ScannerDeviceCreateRequest": {"properties": {"name": {"type": "string", "maxLength": 255, "minLength": 1, "title": "Name"}, "scopes": {"items": {"type": "string"}, "type": "array", "title": "Scopes", "default": ["scan:verify"]}}, "type": "object", "required": ["name"], "title": "ScannerDeviceCreateRequest", "description": "Scanner device creation request."}, "ScannerDeviceCreateResponse": {"properties": {"device_id": {"type": "string", "format": "uuid", "title": "Device Id"}, "api_key": {"type": "string", "title": "Api Key"}, "hmac_secret": {"type": "string", "title": "Hmac Secret"}, "name": {"type": "string", "title": "Name"}, "scopes": {"items": {"type": "string"}, "type": "array", "title": "Scopes"}}, "type": "object", "required": ["device_id", "api_key", "hmac_secret", "name", "scopes"], "title": "ScannerDeviceCreateResponse", "description": "Scanner device creation response (API key returned once)."}, "TokenCreateRequest": {"properties": {"payload_url": {"type": "string", "maxLength": 2048, "title": "Payload Url", "description": "The target URL or URI the QR code will direct to after verification."}, "expires_in_hours": {"anyOf": [{"type": "integer", "maximum": 8760.0, "minimum": 1.0}, {"type": "null"}], "title": "Expires In Hours", "description": "Optional expiration limit in hours from creation."}}, "type": "object", "required": ["payload_url"], "title": "TokenCreateRequest", "description": "Payload details for generating a new SVT."}, "TokenDetailsResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "trace_id": {"type": "string", "title": "Trace Id"}, "payload_url": {"type": "string", "title": "Payload Url"}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "expires_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Expires At"}, "revoked_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Revoked At"}, "revocation_reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Revocation Reason"}}, "type": "object", "required": ["id", "trace_id", "payload_url", "status", "created_at"], "title": "TokenDetailsResponse", "description": "Detailed view of an existing token."}, "TokenRefreshResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "TokenRefreshResponse", "description": "Token refresh response."}, "TokenResponse": {"properties": {"trace_id": {"type": "string", "title": "Trace Id"}, "status": {"type": "string", "title": "Status"}, "expires_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Expires At"}, "qr_code_base64": {"type": "string", "title": "Qr Code Base64", "description": "Base64-encoded PNG image of the standard QR Code containing the SVT."}, "svt_raw": {"type": "string", "title": "Svt Raw", "description": "The raw CBOR + Base64url encoded envelope for manual testing/transport."}}, "type": "object", "required": ["trace_id", "status", "qr_code_base64", "svt_raw"], "title": "TokenResponse", "description": "Response containing SVT generation details and the raw QR image."}, "UserLoginRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "UserLoginRequest", "description": "Login request schema."}, "UserLoginResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "UserLoginResponse", "description": "Login response schema (access token in body)."}, "UserRegisterRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "maxLength": 128, "minLength": 8, "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "UserRegisterRequest", "description": "Registration request schema."}, "UserRegisterResponse": {"properties": {"user_id": {"type": "string", "format": "uuid", "title": "User Id"}, "email": {"type": "string", "title": "Email"}, "role": {"type": "string", "title": "Role"}}, "type": "object", "required": ["user_id", "email", "role"], "title": "UserRegisterResponse", "description": "Registration response schema."}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/api/v1/auth/login"}}}}}}